+++ src/commonMain/kotlin/com/algolia/search/util/internal/Generics.kt
@@ -1,0 +1,4 @@
+package com.algolia.search.util.internal
+
+@Suppress("UNCHECKED_CAST")
+internal fun <T> Any.cast(): T = this as T