+++ src/commonTest/kotlin/serialize/settings/TestSettings.kt
@@ -34,9 +34,11 @@ import com.algolia.search.serialize.KeyAttributesForFaceting
 import com.algolia.search.serialize.KeyAttributesToHighlight
 import com.algolia.search.serialize.KeyAttributesToRetrieve
 import com.algolia.search.serialize.KeyAttributesToSnippet
+import com.algolia.search.serialize.KeyAttributesToTransliterate
 import com.algolia.search.serialize.KeyCamelCaseAttributes
 import com.algolia.search.serialize.KeyCustomNormalization
 import com.algolia.search.serialize.KeyCustomRanking
+import com.algolia.search.serialize.KeyDecompoundQuery
 import com.algolia.search.serialize.KeyDecompoundedAttributes
 import com.algolia.search.serialize.KeyDisableExactOnAttributes
 import com.algolia.search.serialize.KeyDisablePrefixOnAttributes
@@ -66,6 +68,7 @@ import com.algolia.search.serialize.KeyPrimary
 import com.algolia.search.serialize.KeyQueryLanguages
 import com.algolia.search.serialize.KeyQueryType
 import com.algolia.search.serialize.KeyRanking
+import com.algolia.search.serialize.KeyRelevancyStrictness
 import com.algolia.search.serialize.KeyRemoveStopWords
 import com.algolia.search.serialize.KeyRemoveWordsIfNoResults
 import com.algolia.search.serialize.KeyReplaceSynonymsInHighlight
@@ -169,7 +172,10 @@ internal class TestSettings : TestSerializer<Settings>(Settings.serializer()) {
             indexLanguages = listOf(Language.Japanese),
             customNormalization = mapOf(unknown to mapOf(unknown to unknown)),
             enablePersonalization = boolean,
-            attributeCriteriaComputedByMinProximity = boolean
+            attributeCriteriaComputedByMinProximity = boolean,
+            relevancyStrictness = int,
+            decompoundQuery = boolean,
+            attributesToTransliterate = attributes
         ) to buildJsonObject {
             // Attributes
             put(KeySearchableAttributes, attributesJson)
@@ -260,6 +266,9 @@ internal class TestSettings : TestSerializer<Settings>(Settings.serializer()) {
             put(KeyCustomNormalization, buildJsonObject { put(unknown, buildJsonObject { put(unknown, unknown) }) })
             put(KeyEnablePersonalization, boolean)
             put(KeyAttributeCriteriaComputedByMinProximity, boolean)
+            put(KeyRelevancyStrictness, int)
+            put(KeyDecompoundQuery, boolean)
+            put(KeyAttributesToTransliterate, attributesJson)
         }
     )